Advanced SEO: Save custom title format to API
Part of #5834

The prior work in #6467 and #6071 added a UI element to edit custom SEO
title formats inside of Calypso but they did not save the data back
through the API to the servers.

This patch piggy-backs on the settings in `form-seo.jsx` to send changes
setting updates to WordPress.com.

New `fromApi` and `toApi` functions have been created in
`components/seo/meta-title-editor/mappings.js` to foster reusability for
the conversion code to transform between API-native and Calypso-native
formats.

A new `onChange` hook has also been added to the editor so that the SEO
form can be notified of changes. These changes are stateful at the
component level and not at the Redux state level because there remains
considerable work to bring the Redux tree to the point where it need to
be to make this smooth.

Included in these changes are some reformattings of the **lodash**
imports.

**Remaining Questions**
 - Should we be moving the `mappings.js` file into a place better suited
   for it? Maybe into `client/state/sites/` I'm not sure if that would
   be better or if its current home isn't the best place for it to be.
   It's currently only marginally referenced outside of the component.